deserializeElementsFromNode will now put substitution elements in a single array.
So if you have a type Base with substitutions A and B, deserializing A1, B1, A2, B2 will result in a Base array with A1, B1, A2, B2. Before you'd get back a property A with A2 and a property B with B2. A1 and B1 would get overwritten by A2 and B2.